PRE REFACTOR not have products and coins in the machine
Created a new product object as think it'd be easier as has price
and is a seperate thing to the vending machine. So have created that
and now have a printout for products and the count of them in the
machine... but not added the object to tests yet, will refactor the tests
as working, but tests failing as have string instead of object as
arguments.